whitespace normalization per style guide
I'm about to make some other changes to env.rb and since my editor
is removing trailing whitespace automatically (as all of yours should!)
it is crufting up my other diffs... so this is going in stand alone.
You all should do the same if you see that happen too.

Change-Id: I2387ffa5ea65577afe296e2905f240545d1e7b13