provides a directory structure where we can add functional tests
Functional tests are where we do whitebox testing against core internals.
For starters, we mirror the directory structure of development. We currently have:
-cloudcontroller
-dea
-health_manager
-router
-stager
-staging
BVTs are essentially blackbox testing where we have a targeted cloud and run
a set of apps with operations against it. Functionals are where we
launch kernel processes directly and manipulate internal classes.

Change-Id: I145769a82dbb3e931bae2c03cb6318a5d62cf319